+++ native/src/core/cj_object.cpp
@@ -2,7 +2,7 @@
 
 extern "C"
 {
-    long nativeObjectCreate(long parentPtr)
+    uintptr_t nativeObjectCreate(uintptr_t parentPtr)
     {
         CjObject *obj;
         if (parentPtr == 0l)
@@ -14,9 +14,9 @@ extern "C"
             QObject *parent = reinterpret_cast<QObject *>(static_cast<uintptr_t>(parentPtr));
             obj = new CjObject(parent);
         }
-        return reinterpret_cast<long>(obj);
+        return reinterpret_cast<uintptr_t>(obj);
     }
-    void nativeObjectDelete(long ptr)
+    void nativeObjectDelete(uintptr_t ptr)
     {
         CjObject *instance = reinterpret_cast<CjObject *>(static_cast<uintptr_t>(ptr));
         delete instance;
