+++ native/src/core/event.cpp
@@ -2,42 +2,42 @@
 
 extern "C"
 {
-    long nativeEventCreate(int type)
+    uintptr_t nativeEventCreate(int type)
     {
         QEvent *event = new QEvent(QEvent::Type(type));
-        return reinterpret_cast<long>(event);
+        return reinterpret_cast<uintptr_t>(event);
     }
-    void nativeEventDelete(long ptr)
+    void nativeEventDelete(uintptr_t ptr)
     {
         QEvent *instance = reinterpret_cast<QEvent *>(static_cast<uintptr_t>(ptr));
         delete instance;
     }
-    void nativeEventAccept(long ptr)
+    void nativeEventAccept(uintptr_t ptr)
     {
         QEvent *instance = reinterpret_cast<QEvent *>(static_cast<uintptr_t>(ptr));
         return instance->accept();
     }
-    void nativeEventIgnore(long ptr)
+    void nativeEventIgnore(uintptr_t ptr)
     {
         QEvent *instance = reinterpret_cast<QEvent *>(static_cast<uintptr_t>(ptr));
         return instance->ignore();
     }
-    bool nativeEventIsAccepted(long ptr)
+    bool nativeEventIsAccepted(uintptr_t ptr)
     {
         QEvent *instance = reinterpret_cast<QEvent *>(static_cast<uintptr_t>(ptr));
         return instance->isAccepted();
     }
-    void nativeEventSetAccepted(long ptr, bool accepted)
+    void nativeEventSetAccepted(uintptr_t ptr, bool accepted)
     {
         QEvent *instance = reinterpret_cast<QEvent *>(static_cast<uintptr_t>(ptr));
         return instance->setAccepted(accepted);
     }
-    bool nativeEventSpontaneous(long ptr)
+    bool nativeEventSpontaneous(uintptr_t ptr)
     {
         QEvent *instance = reinterpret_cast<QEvent *>(static_cast<uintptr_t>(ptr));
         return instance->spontaneous();
     }
-    int nativeEventType(long ptr)
+    int nativeEventType(uintptr_t ptr)
     {
         QEvent *instance = reinterpret_cast<QEvent *>(static_cast<uintptr_t>(ptr));
         return instance->type();
