+++ native/src/core/point.cpp
@@ -2,38 +2,38 @@
 
 extern "C"
 {
-    long nativePointCreateWithXY(int xpos, int ypos)
+    uintptr_t nativePointCreateWithXY(int xpos, int ypos)
     {
         QPoint *point = new QPoint(xpos, ypos);
-        return reinterpret_cast<long>(point);
+        return reinterpret_cast<uintptr_t>(point);
     }
-    long nativePointCreate()
+    uintptr_t nativePointCreate()
     {
         QPoint *point = new QPoint();
-        return reinterpret_cast<long>(point);
+        return reinterpret_cast<uintptr_t>(point);
     }
 
-    void nativePointDelete(long ptr)
+    void nativePointDelete(uintptr_t ptr)
     {
         QPoint *instance = reinterpret_cast<QPoint *>(static_cast<uintptr_t>(ptr));
         delete instance;
     }
-    int nativePointRx(long ptr)
+    int nativePointRx(uintptr_t ptr)
     {
         QPoint *instance = reinterpret_cast<QPoint *>(static_cast<uintptr_t>(ptr));
         return instance->rx();
     }
-    int nativePointRy(long ptr)
+    int nativePointRy(uintptr_t ptr)
     {
         QPoint *instance = reinterpret_cast<QPoint *>(static_cast<uintptr_t>(ptr));
         return instance->ry();
     }
-    int nativePointX(long ptr)
+    int nativePointX(uintptr_t ptr)
     {
         QPoint *instance = reinterpret_cast<QPoint *>(static_cast<uintptr_t>(ptr));
         return instance->x();
     }
-    int nativePointY(long ptr)
+    int nativePointY(uintptr_t ptr)
     {
         QPoint *instance = reinterpret_cast<QPoint *>(static_cast<uintptr_t>(ptr));
         return instance->y();
