+++ native/src/core/rect.cpp
@@ -2,12 +2,12 @@
 
 extern "C"
 {
-    long nativeRectCreate(int x, int y, int width, int height)
+    uintptr_t nativeRectCreate(int x, int y, int width, int height)
     {
         QRect *rect = new QRect(x, y, width, height);
-        return reinterpret_cast<long>(rect);
+        return reinterpret_cast<uintptr_t>(rect);
     }
-    void nativeRectDelete(long ptr)
+    void nativeRectDelete(uintptr_t ptr)
     {
         QRect *instance = reinterpret_cast<QRect *>(static_cast<uintptr_t>(ptr));
         delete instance;
