+++ native/src/core/size.cpp
@@ -2,47 +2,47 @@
 
 extern "C"
 {
-    long nativeSizeCreate()
+    uintptr_t nativeSizeCreate()
     {
         QSize *size = new QSize();
-        return reinterpret_cast<long>(size);
+        return reinterpret_cast<uintptr_t>(size);
     }
-    long nativeSizeCreateWithWidth(int width, int height)
+    uintptr_t nativeSizeCreateWithWidth(int width, int height)
     {
         QSize *size = new QSize(width, height);
-        return reinterpret_cast<long>(size);
+        return reinterpret_cast<uintptr_t>(size);
     }
-    void nativeSizeDelete(long ptr)
+    void nativeSizeDelete(uintptr_t ptr)
     {
         QSize *instance = reinterpret_cast<QSize *>(static_cast<uintptr_t>(ptr));
         delete instance;
     }
 
-    int nativeSizeWidth(long ptr)
+    int nativeSizeWidth(uintptr_t ptr)
     {
         QSize *instance = reinterpret_cast<QSize *>(static_cast<uintptr_t>(ptr));
         return instance->width();
     }
 
-    int nativeSizeHeight(long ptr)
+    int nativeSizeHeight(uintptr_t ptr)
     {
         QSize *instance = reinterpret_cast<QSize *>(static_cast<uintptr_t>(ptr));
         return instance->height();
     }
 
-    bool nativeSizeIsNull(long ptr)
+    bool nativeSizeIsNull(uintptr_t ptr)
     {
         QSize *instance = reinterpret_cast<QSize *>(static_cast<uintptr_t>(ptr));
         return instance->isNull();
     }
 
-    bool nativeSizeIsEmpty(long ptr)
+    bool nativeSizeIsEmpty(uintptr_t ptr)
     {
         QSize *instance = reinterpret_cast<QSize *>(static_cast<uintptr_t>(ptr));
         return instance->isEmpty();
     }
 
-    bool nativeSizeIsValid(long ptr)
+    bool nativeSizeIsValid(uintptr_t ptr)
     {
         QSize *instance = reinterpret_cast<QSize *>(static_cast<uintptr_t>(ptr));
         return instance->isValid();
